requirements.txt
fastapi
uvicorn
python-multipart
llama-cpp-python
matplotlib